new vinylMode param
Add a new parameter to params, `vinylmode`, which if present and
true changes the plugin so that it expects to get images piped in,
and in turn pipes icon image files out. This also adds an optional
htmlCodeCallback that is called with a string of html code
corresponding to the streamed icon files.